requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1
feedparser==6.0.11
httpx==0.27.2
beautifulsoup4==4.12.3
